+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Fixed handle reconnection timeout error 
 * Removed check the node is available for query and table service sessions 
 * Refactored the `balancers.PreferLocations()` function - it is a clean/pure function
 * Added experimental `balancers.WithNodeID()` context modifier for define per request the YDB endpoint by NodeID